Refactor background scripts
