+++ src/background/index.js
@@ -1,7 +1,26 @@
-import * as documentModule from "./document";
-import * as startModule from "./start";
+import { getStorage } from "../common/storage";
+import { defaults } from "../common/settings";
+
+const getElements = () => {
+    return document.querySelectorAll(".ghx-heading .ghx-info > span");
+}
+
+function start() {
+    getStorage("Jira-Status-Colours-Configuration", (result) => {
+        const savedConfiguration = result || defaults;
+        const elements = getElements();
+
+        for (let i = 0; i < elements.length; i++) {
+            const configuration = savedConfiguration[elements[i].innerText.toUpperCase()];
+            
+            if(configuration) {
+                elements[i].style.backgroundColor = configuration.backgroundColour;
+                elements[i].style.color = configuration.textColour;
+            }
+        }
+    });    
+}
 
 setTimeout(function() {
-    documentModule.initializeDocument(document);
-    startModule.start();
+    start();
 }, 500);
